Refactor storage to storage node
