+++ plugins/alerting/pkg/alerting/util.go
@@ -1,7 +1,6 @@
 package alerting
 
 import (
-	"context"
 	"time"
 
 	prommodel "github.com/prometheus/common/model"
@@ -43,8 +42,4 @@ func NewCortexAlertingRule(alertId string, interval *time.Duration, rule metrics
 		Interval: promInterval,
 		Rules:    []rulefmt.Rule{*actualRuleFmt},
 	}, nil
-}
-
-func setPluginHandlerTimeout(ctx context.Context, dur time.Duration) (context.Context, context.CancelFunc) {
-	return context.WithTimeout(ctx, dur)
 }
